Rearchitecture JacLLM Call
