+++ jaclang/plugin/spec.py
@@ -214,25 +214,6 @@ class JacFeatureSpec:
 
     @staticmethod
     @hookspec(firstresult=True)
-    def get_semstr_type(
-        file_loc: str, scope: str, attr: str, return_semstr: bool
-    ) -> Optional[str]:
-        """Jac's get_semstr_type stmt feature."""
-        raise NotImplementedError
-
-    @staticmethod
-    @hookspec(firstresult=True)
-    def obj_scope(file_loc: str, attr: str) -> str:
-        """Jac's get_semstr_type feature."""
-        raise NotImplementedError
-
-    @staticmethod
-    def get_sem_type(file_loc: str, attr: str) -> tuple[str | None, str | None]:
-        """Jac's get_semstr_type feature."""
-        raise NotImplementedError
-
-    @staticmethod
-    @hookspec(firstresult=True)
     def with_llm(
         file_loc: str,
         model: Any,  # noqa: ANN401
@@ -264,6 +245,29 @@ class JacBuiltin:
         dot_file: Optional[str],
     ) -> str:
         """Print the dot graph."""
+        raise NotImplementedError
+
+
+class JacLLM:
+    """Jac LLM."""
+
+    @staticmethod
+    @hookspec(firstresult=True)
+    def get_semstr_type(
+        file_loc: str, scope: str, attr: str, return_semstr: bool
+    ) -> Optional[str]:
+        """Jac's get_semstr_type stmt feature."""
+        raise NotImplementedError
+
+    @staticmethod
+    @hookspec(firstresult=True)
+    def obj_scope(file_loc: str, attr: str) -> str:
+        """Jac's get_semstr_type feature."""
+        raise NotImplementedError
+
+    @staticmethod
+    def get_sem_type(file_loc: str, attr: str) -> tuple[str | None, str | None]:
+        """Jac's get_semstr_type feature."""
         raise NotImplementedError
